Extract ExternalLink component with styling and icon
Replaces the duplicated link style + ExternalLinkIcon pattern across
MoviePage, CinemaInfoBar, and About with a single ExternalLink component
that encapsulates the anchor tag, styling, and icon.
